Created MenuDisplayHelper and refactoring Menu further
